First version working with both Django 1.2.x and 1.3.x.
There are some snafus though:

As I found out, you can't reliably convert the django date formats to
Datepicker formats. So this commit introduces these two new settings:

    FILTRATE = {
        'datepicker_region': 'da',
        'datepicker_date_format': 'dd.mm.y',
    }

Check out the Datepicker documentation to see how to use them.

Django 1.2.4 introduces restrictions which lookups can be queried
in the url, so at the moment the end user are responsible for
checking for those, as in this example:

    class CaseAdmin(admin.ModelAdmin):
	list_filter = ['client']

	class Media():
		pass

	def lookup_allowed(self, key, *args, **kwargs):
		if 'client__start_date' in key:
			return True
		else:
			return super(CaseAdmin, self).lookup_allowed(key, *args, **kwargs)

Time and my Python meta-fu is running out, and I couldn't fix it
so its not neccessary to define an empty Media() class. See the
example above.